import fixes and moved factory.py
